#1120 :truck: rename `wahlWahlartEnum`
