+++ wls-gui-wahllokalsystem/src/types/Task.ts
@@ -1,4 +1,4 @@
-import type { WahlWahlartEnum } from "@/types/wahl/wahlWahlartEnum.ts";
+import type { WahlWahlartEnum } from "@/types/wahl/WahlWahlartEnum.ts";
 import type { WahlbezirksArtEnum } from "@/types/wahlbezirksArtEnum.ts";
 
 export interface Task {
