+++ wls-gui-wahllokalsystem/src/types/wahl/Wahl.ts
@@ -1,6 +1,6 @@
 import type { Farbe } from "@/types/wahl/Farbe.ts";
 
-import { WahlWahlartEnum } from "@/types/wahl/wahlWahlartEnum.ts";
+import { WahlWahlartEnum } from "@/types/wahl/WahlWahlartEnum.ts";
 
 export interface Wahl {
   wahlID: string;
