+++ wls-gui-wahllokalsystem/src/types/wahl/WahlWahlartEnum.ts
