+++ wls-gui-wahllokalsystem/tests/composables/wahl/wahlMapper.spec.ts
@@ -6,7 +6,7 @@ import { describe, expect, it } from "vitest";
 
 import { WahlDTOWahlartEnum } from "@/api/wls-clients/generated-basisdaten-api";
 import { useWahlMapper } from "@/composables/wahl/wahlMapper.ts";
-import { WahlWahlartEnum } from "@/types/wahl/wahlWahlartEnum.ts";
+import { WahlWahlartEnum } from "@/types/wahl/WahlWahlartEnum.ts";
 
 describe("WahlMapper", () => {
   const { toModel } = useWahlMapper();
